requirements.txt
langchain==0.0.149
python-dotenv==1.0.0
supabase==1.0.3
tiktoken==0.3.3
openai
python-dotenv
vecs
